+++ src/util/Constants.ts
@@ -1,0 +1,5 @@
+const sasToken =
+    "?sv=2020-08-04&ss=b&srt=sco&sp=rwdlacitfx&se=2022-09-30T20:46:30Z&st=2021-11-30T13:46:30Z&spr=https&sig=e3MgX4FkExWFc5LWFv5RGZdLdyzSMy3ZSSX76%2BPzGFs%3D";
+
+const fileNamePostfix  = 5;
+export {sasToken, fileNamePostfix};
