+++ src/util/saveExperienceToCloud.ts
@@ -1,7 +1,6 @@
 import { BlobServiceClient, ContainerClient } from "@azure/storage-blob";
+import {sasToken} from "./Constants";
 
-const sasToken =
-  "?sv=2020-08-04&ss=b&srt=sco&sp=rwdlacitfx&se=2022-09-30T20:46:30Z&st=2021-11-30T13:46:30Z&spr=https&sig=e3MgX4FkExWFc5LWFv5RGZdLdyzSMy3ZSSX76%2BPzGFs%3D";
 
 const saveExperienceToCloud = async (
   userId: string,
